double for transaction response

+++ lib/src/moyasar_transfer/domain/entity/transaction_response.dart
@@ -1,8 +1,6 @@
-
-
 class TransactionResponse {
-  final int fee;
-  final int amount;
+  final double fee;
+  final double amount;
 
   TransactionResponse({
     required this.fee,
